requirements.txt
langgraph
langchain
langchain_google_genai
python-dotenv
pandas
faiss-cpu
kagglehub
requests
duckduckgo-search
spacy